+++ src/app/dishes/[id]/dish-description.tsx
@@ -1,16 +1,14 @@
 import styles from '@/styles/app/dishes/[id]/page.module.scss';
-import { DetailedDish } from '@/src/types/api.types';
 
 interface DishDescriptionProps {
-    dish: DetailedDish;
-    description?: string;
+    description: string;
 }
 
-export function DishDescription({ dish, description }: DishDescriptionProps) {
+export function DishDescription({ description }: DishDescriptionProps) {
     return (
         <div className={styles['dish-description']}>
             <h5>Opis:</h5>
-            <p dangerouslySetInnerHTML={{ __html: description && description.length > 0 ? description : dish.description }} />
+            <p dangerouslySetInnerHTML={{ __html: description?.length > 0 ? description : 'Brak dostarczonego opisu.' }} />
         </div>
     );
 }