+++ src/components/common/links/edit-link.tsx
@@ -3,7 +3,7 @@
 import EditIcon from '@mui/icons-material/Edit';
 import styles from '@/styles/components/common/buttons/add-button.module.scss';
 import { useRouter } from 'next/navigation';
-import { PagePathname } from '@/src/constants/strings';
+import { PagePathname } from '@/src/constants/strings.constants';
 
 interface EditLinkProps {
     link: PagePathname;